+++ src/components/BsNotification/index.js
@@ -5,8 +5,8 @@ export default Vue => {
     const myComponent = Vue.extend(BsNotification);
     const component = new myComponent().$mount()
 
-    document.body.appendChild(component.$el);
     registerPrototype(Vue);
+    setTimeout(() => document.body.appendChild(component.$el), 100);
     Vue.prototype.$VueMdb.$notification = component;
     Vue.prototype.$notification = Vue.prototype.$VueMdb.$notification;
     Vue.component(BsNotification.name, BsNotification);